React-landing

+++ src/main/java/juandaza/movierenter/Repository/RepoCliente.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import juandaza.movierenter.Model.Cliente;
 
-public interface RepoCliente extends CrudRepository<Cliente, Integer> {
+public interface RepoCliente extends CrudRepository<Cliente, Long> {
 
 }
